+++ crates/acts/src/store/data/event.rs
@@ -10,7 +10,7 @@ pub struct Event {
     pub id: String,
     pub name: String,
     pub mid: String,
-    pub ver: i32,
+    pub ver: String,
 
     pub uses: String,
     pub params: String,
@@ -26,12 +26,12 @@ impl DbCollectionIden for Event {
 }
 
 impl Event {
-    pub fn from_act(act: &Act, mid: &str, ver: i32, event_id: &str) -> Result<Self> {
+    pub fn from_act(act: &Act, mid: &str, ver: &str, event_id: &str) -> Result<Self> {
         Ok(Self {
             id: event_id.to_string(),
             name: act.name.to_string(),
             mid: mid.to_string(),
-            ver,
+            ver: ver.to_string(),
             uses: act.uses.clone(),
             params: serde_json::to_string(&act.params).map_err(|err| {
                 ActError::Convert(format!("failed to convert params to string: {err}"))