+++ crates/acts/src/store/tests/mem.rs
@@ -22,7 +22,8 @@ async fn store_mem_model_create() {
     let model = Model {
         id: utils::longid(),
         name: "test".to_string(),
-        ver: 1,
+        desc: "test desc".to_string(),
+        ver: "0.1.0".to_string(),
         size: 1245,
         create_time: 3333,
         update_time: 0,
@@ -40,7 +41,8 @@ async fn store_mem_model_find() {
     let model = Model {
         id: mid.clone(),
         name: "test".to_string(),
-        ver: 1,
+        desc: "test desc".to_string(),
+        ver: "0.1.0".to_string(),
         size: 1245,
         create_time: 3333,
         data: "{}".to_string(),
@@ -59,7 +61,8 @@ async fn store_mem_model_query_id() {
         let model = Model {
             id: utils::longid(),
             name: "test_model".to_string(),
-            ver: 1,
+            desc: "test desc".to_string(),
+            ver: "0.1.0".to_string(),
             size: 1245,
             create_time: 3333,
             update_time: 0,
@@ -83,7 +86,8 @@ async fn store_mem_model_query_match_or() {
         let model = Model {
             id: utils::longid(),
             name: format!("test_model {i}"),
-            ver: 1,
+            desc: "test desc".to_string(),
+            ver: "0.1.0".to_string(),
             size: 1000,
             create_time: 3333,
             update_time: 0,
@@ -112,7 +116,8 @@ async fn store_mem_model_query_match_and() {
         let model = Model {
             id: utils::longid(),
             name: format!("test_model {i}"),
-            ver: 1,
+            desc: "test desc".to_string(),
+            ver: "0.1.0".to_string(),
             size: 2000,
             create_time: 3333,
             update_time: 0,
@@ -141,7 +146,8 @@ async fn store_mem_model_update() {
     let mut model = Model {
         id: utils::longid(),
         name: "test".to_string(),
-        ver: 1,
+        desc: "test desc".to_string(),
+        ver: "0.1.0".to_string(),
         size: 1245,
         create_time: 3333,
         update_time: 0,
@@ -150,7 +156,7 @@ async fn store_mem_model_update() {
     };
     store.models().create(&model).unwrap();
 
-    model.ver = 3;
+    model.ver = "0.2.0".to_string();
     model.update_time = 1;
     store.models().update(&model).unwrap();
 
@@ -165,7 +171,8 @@ async fn store_mem_model_delete() {
     let model = Model {
         id: utils::longid(),
         name: "test".to_string(),
-        ver: 1,
+        desc: "test desc".to_string(),
+        ver: "0.1.0".to_string(),
         size: 1245,
         create_time: 3333,
         update_time: 0,
@@ -1022,7 +1029,7 @@ async fn store_mem_event_create() {
         id,
         name: "name".to_string(),
         mid: "mid".to_string(),
-        ver: 1,
+        ver: "0.1.0".to_string(),
         uses: "acts.event.manual".to_string(),
         params: "".to_string(),
         create_time: utils::time::time_millis(),
@@ -1043,7 +1050,7 @@ async fn store_mem_event_query_id() {
         id,
         name: "name".to_string(),
         mid: "mid".to_string(),
-        ver: 1,
+        ver: "0.1.0".to_string(),
         uses: "acts.event.manual".to_string(),
         params: "".to_string(),
         create_time: 0,
@@ -1065,7 +1072,7 @@ async fn store_mem_event_query_match_or() {
             id,
             name: format!("name {idx}"),
             mid: "mid1".to_string(),
-            ver: 1,
+            ver: "0.1.0".to_string(),
             uses: "acts.event.manual".to_string(),
             params: "".to_string(),
             create_time: 0,
@@ -1096,7 +1103,7 @@ async fn store_mem_event_query_match_and() {
             id,
             name: format!("name {idx}"),
             mid: "mid2".to_string(),
-            ver: 1,
+            ver: "0.1.0".to_string(),
             uses: "acts.event.manual".to_string(),
             params: "".to_string(),
             create_time: 0,
@@ -1126,7 +1133,7 @@ async fn store_mem_event_update() {
         id,
         name: "name".to_string(),
         mid: "mid".to_string(),
-        ver: 1,
+        ver: "0.1.0".to_string(),
         uses: "acts.event.manual".to_string(),
         params: "".to_string(),
         create_time: 0,
@@ -1155,7 +1162,7 @@ async fn store_mem_event_remove() {
         id,
         name: "name".to_string(),
         mid: "mid".to_string(),
-        ver: 1,
+        ver: "0.1.0".to_string(),
         uses: "acts.event.manual".to_string(),
         params: "".to_string(),
         create_time: 0,